// src/enums/user.enum.ts
enum INITIAL {
  MRS = 'MRS',
  MR = 'MR',
  MISS = "MISS",
  MS = "MS",
  MASTER = "MASTER",
  BABY = "BABY",
  MST = "MST",
  DR = "DR"
}

enum RELATION {
  FAMILY = "FAMILY",
  FRIEND = "FRIEND",
  BUSINESS = "BUSINESS",
  OTHER = "OTHER"
}

enum MARITAL_STATUS {
  SINGLE = "SINGLE",
  MARRIED = "MARRIED"
}

enum GENDER {
  MALE = 'MALE',
  FEMALE = "FEMALE"
}

enum TYPE {
  USER = 'USER',
  FAMILY_MEMBER = "FAMILY_MEMBER",
  STAFF_MEMBER = "STAFF_MEMBER"
}

enum RESIDENTORVISITOR {
  RESIDENT = "RESIDENT",
  VISITOR = "VISITOR"
}

enum OCCUPATION {
  GOVERNMENT = "GOVERNMENT",
  PRIVATE = "PRIVATE",
  HOUSEWIFE = "HOUSEWIFE",
  RETIRED = "RETIRED",
  STUDENT = "STUDENT",
  TEACHER = "TEACHER",
  UNEMPLOYED = "UNEMPLOYED",
  CHILD = "CHILD",
  LABORCAMP = "LABORCAMP"
}

export {
  GENDER,
  MARITAL_STATUS,
  RELATION,
  INITIAL,
  TYPE,
  RESIDENTORVISITOR,
  OCCUPATION
}